Check movie today has been show poster movie

+++ app/src/main/java/ysn/temanbioskop/internal/model/bioskop/moviedb/search/Result.java
@@ -52,7 +52,7 @@ public class Result {
     private Boolean video;
     @SerializedName("vote_average")
     @Expose
-    private Integer voteAverage;
+    private float voteAverage;
 
     public String getPosterPath() {
         return posterPath;
@@ -158,11 +158,11 @@ public class Result {
         this.video = video;
     }
 
-    public Integer getVoteAverage() {
+    public float getVoteAverage() {
         return voteAverage;
     }
 
-    public void setVoteAverage(Integer voteAverage) {
+    public void setVoteAverage(float voteAverage) {
         this.voteAverage = voteAverage;
     }
